% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_interval}
\alias{add_interval}
\alias{add_interval_baseline}
\title{Add intervals}
\usage{
add_interval(data, name = interval)

add_interval_baseline(data, name = interval_baseline)
}
\arguments{
\item{data}{data extracted from the NOAS}

\item{name}{unquoted name to give the new column}
}
\value{
data frame with one extra column
}
\description{
Since we have longitudinal data,
intervals for subjects is a good  way
to get a better idea of the time scale
of the data. These functions add extra
columns to your data.
}
\details{
\itemize{
\item{add_interval}{ - add interval since last visit - default col: interval}
\item{add_interval_baaseline}{ - add interval since first visit- default col: baseline}
}
}
\examples{
# attach built-in noas example data to test
dt <- noas_example

add_interval(dt)
add_interval_baseline(dt)

library(dplyr)
dt \%>\%
  add_interval() \%>\%
  add_interval_baseline()

# Change the default column names
dt \%>\%
  add_interval(name = intv) \%>\%
  add_interval_baseline(name = bsl_intv)
}
