% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic-funcs.R
\name{add_pgs}
\alias{add_pgs}
\title{Add PGS data to the MOAS}
\usage{
add_pgs(MOAS, pgs = c("AD", "AD_Jansen"), s_levels = c("S1", "S7",
  "S11"),
  pgs_path = "~/LCBC/Projects/Cross_projects/Genetics/PGS/PGS_20190618/PGS_wAPOE/",
  genetic_match_file = "~/LCBC/Projects/Cross_projects/MOAS/data-raw/DNA/gID_MOAS_match.tsv",
  include_cnt = FALSE, include_genetics_debug = FALSE)
}
\arguments{
\item{MOAS}{data.frame of the full MOAS or MOAS sub-set}

\item{pgs}{character vector of the PGS wanted}

\item{s_levels}{character vector of the significance levels wanted}

\item{pgs_path}{path to the directory where the PGS's are stored}

\item{genetic_match_file}{path to the file containing the MOAS-genetics
matching and debugging information}

\item{include_cnt}{logical, whether to keep SNP count information}

\item{include_genetics_debug}{logical, whether to keep all columns
in the genetic_match_file in the final output}
}
\value{
a tibble / data.frame
}
\description{
This function calls on [\code{get_pgs}] to read in wanted
PGS at specified significance levels, and adds that data
to the MOAS-type data provided.
}
\examples{
\dontrun{

# Here we assume you know the path to the LCBC
# lagringshotell, you can substitute "~" in the
# paths with the path to the lagringshotell

add_pgs(pgs = c("AD", "EduYears_2016", "Depression_Nagel2018"),
    s_levels = c("S1", "S7", "S11"),
    pgs_path = "~/LCBC/Projects/Cross_projects/Genetics/PGS/PGS_20190618/PGS_wAPOE/",
    genetic_match_file = "~/LCBC/Projects/Cross_projects/MOAS/data-raw/DNA/gID_MOAS_match.tsv",
)
}
}
