% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_timepoint}
\alias{add_timepoint}
\title{Add timepoint to data}
\usage{
add_timepoint(data, var_name = timepoint)
}
\arguments{
\item{data}{data extracted from the NOAS}

\item{var_name}{unquoted name to give the new column}
}
\value{
data frame with extra column with timepoint
}
\description{
Calculated sequential timepoint
for participants along the 'age'
column.
}
