% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{anonymize_moas}
\alias{anonymize_moas}
\title{Anonymise the MOAS}
\usage{
anonymize_moas(data, scramble_ids = TRUE)
}
\arguments{
\item{data}{The MOAS or a MOAS generated file.}

\item{scramble_ids}{make IDs anonymous}
}
\value{
tibble
}
\description{
The function calls both [\code{deselect_sensitive}] and [\code{mutate_sensitive}]
to remove sensitive columns, and alter data in the MOAS to have a less
sensitive nature.
}
