% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_date}
\alias{as_date}
\title{Coerce NOAS date into R date}
\usage{
as_date(dates)
}
\arguments{
\item{dates}{character of dates (yyyy-mm-dd)}
}
\value{
vector of R dates
}
\description{
Dates from the NOAS always come
in a specific format (yyyy-mm-dd).
This function is for convenience in
turning NOAS dates into R-dates easily.
}
\examples{
dates <- c("2016-03-01", "2002-12-31")
dates <- as.Date(dates, format="\%Y-\%m-\%d")
dates
as.numeric(dates)
}
