% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseCols}
\alias{baseCols}
\title{MOAS base columns}
\format{An object of class \code{character} of length 12.}
\usage{
data(baseCols)
}
\description{
a string vector with the most basic columns in the MOAS,
that are usually required for MOAS-functions to work.
}
\examples{
data(baseCols)
}
\keyword{datasets}
