% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ages.R
\name{calc_ages}
\alias{calc_ages}
\title{Calculates ages based on birthdates and test/MRI dates, and
calculated subject timepoints etc.}
\usage{
calc_ages(data)
}
\arguments{
\item{data}{The MOAS or a MOAS generated file.}
}
\value{
a data frame with added/replaced age and timepoint variables.
}
\description{
\code{calc_ages} Creates a dataframe where age and subject timepoint
have been added to the data. Both MRI and Test age are calculated,
and "Age" is the mean between these two. If only of of MRI or Test
age are present, this one is used for Age. Also calculates intervals
between test and MRI, baseline to every subsequent testing, and
intervall between timepoints.
}
\examples{
calc_ages(Data)


}
