% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hour.R
\name{calc_hour}
\alias{calc_hour}
\title{Calculates time in decimal hours from HH:MM format.}
\usage{
calc_hour(data)
}
\arguments{
\item{data}{String or time data in 'HH:MM' format}
}
\value{
a numeric vector of decimal hours
}
\description{
\code{calc_hour} calculates time in decimal hours from HH:MM format.
}
\examples{
 calc_hour(data)

}
