% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tp.R
\name{check_tp}
\alias{check_tp}
\title{Check subject timepoint}
\usage{
check_tp(ID = NULL,
  MOAS = "//lagringshotell/sv-psi/LCBC/Projects/Cross_projects/MOAS/Data/MOAS.RData")
}
\arguments{
\item{ID}{7-digit code, CrossProject_ID}

\item{MOAS}{A string with the path to, or the pre-loaded,
MOAS data.frame}
}
\value{
a data.frame with information on the ID
}
\description{
This function is intended for the RA's of LCBC to lookup,
using the MOAS, hich timepoint should be the next allocated
to a participant.
}
