% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_DF.R
\name{clean_DF}
\alias{clean_DF}
\title{Cleans data from SPSS into a data.frame}
\usage{
clean_DF(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
a cleaned data.frame
}
\description{
\code{clean_DF} cleans SPSS imported data of it's many
labels and attributed, for easier reading and handling in R
}
\examples{
\dontrun{
clean_DF(data)
}
}
