% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_chars.R
\name{count_chars}
\alias{count_chars}
\title{Utility function to count the number of occurances
of a specific character in a string}
\usage{
count_chars(char, s)
}
\arguments{
\item{char}{Character to count.}

\item{s}{string vector}
}
\value{
vector of integers
}
\description{
\code{count_chars} counts the number of times a
character occurs in a string
}
\examples{
count_chars('h', 'How many times is 'h' in this string?')


}
