% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{MOAS mock example data}
\format{An object of class \code{data.frame} with 1383 rows and 244 columns.}
\usage{
data(data)
}
\description{
A mock example of the LCBC data
}
\examples{
data(data)
}
\keyword{datasets}
