% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs-funcs.R
\name{dbs_get}
\alias{dbs_get}
\title{Get DBS information from a file}
\usage{
dbs_get(file_path, match_file = "data-raw/Blood_data/DBS_MOAS_match.tsv",
  debug = FALSE)
}
\arguments{
\item{file_path}{path to DBS excel-file}

\item{match_file}{path to the match-file for DBS}

\item{debug}{logical to keep debug information (default: FALSE)}
}
\value{
tibble of DBS data
}
\description{
Get DBS information from a file
}
\examples{
\dontrun{
<<<<<<< HEAD
dbs_get(file_path = "path/to/file"),
       match_file = "path/to/MOAS/data-raw/Blood_data/DBS_MOAS_match.tsv")

dbs_get(file_path = "path/to/file"),
=======
dbs_get(file_path = "path/to/file",
       match_file = "path/to/MOAS/data-raw/Blood_data/DBS_MOAS_match.tsv")

dbs_get(file_path = "path/to/file",
>>>>>>> 4be24afb0e6488045bb28bda464ac6bf3885ec2c
       match_file = "path/to/MOAS/data-raw/Blood_data/DBS_MOAS_match.tsv",
       debug = TRUE)
 }
}
\seealso{
Other MOAS get-functions: \code{\link{epigen_get}},
  \code{\link{pgs_get}}
}
\concept{MOAS get-functions}
\concept{dbs-functions}
