% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{deselect_sensitive}
\alias{deselect_sensitive}
\title{Deselect sensitive columns}
\usage{
deselect_sensitive(data)
}
\value{
tibble
}
\description{
This function will remove all the columns from
a MOAS-like data.frame that are known to include
possible sensitive information.
}
\examples{
dt <- data.frame(ID = 1:3,
    Birth_Date = c("1997-01-12", "1984-01-01", "1953-09-16"),
   Comment = c("", "Has leukemia", "is on diazepam"),
   CVLT_A = c(19, 25, 29),
   stringsAsFactors = FALSE)
deselect_sensitive(dt)
}
