% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-paradigm.R
\name{docs_paradigm_order_chapter}
\alias{docs_paradigm_order_chapter}
\title{Order the chapters of the Paradigm documentation}
\usage{
docs_paradigm_order_chapter(path = ".")
}
\arguments{
\item{path}{path to Paradigm_documentation folder}
}
\description{
Over time, adding new chapters to the documentation
can be tedious as numbering will change. This
functions makes that easy, by ordering and numbering
the chapters by `type` and the alphabetically.
}
