% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-paradigm.R
\name{docs_paradigm_page}
\alias{docs_paradigm_page}
\title{Create new chapter for the Paradigm documentation}
\usage{
docs_paradigm_page(name = NULL, type = NULL, path = ".", bib = TRUE)
}
\arguments{
\item{name}{acronym of chapter}

\item{type}{type of chapter (one of "rt", "sp", "fmri", "erp")}

\item{path}{path to Paradigm documentation}

\item{bib}{logical, initiate bibTex file or not}
}
\description{
Create new chapter for the Paradigm documentation
}
