% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_project.R
\name{docs_project_page}
\alias{docs_project_page}
\title{Create new chapter for the Project documentation}
\usage{
docs_project_page(project_acronym = NULL, project_number = NULL,
  path = ".")
}
\arguments{
\item{project_acronym}{acronym of chapter}

\item{project_number}{type of chapter (one of "rt",
"sp", "fmri", "erp")}

\item{path}{path to Paradigm documentation}
}
\description{
Create new chapter for the Project documentation
}
