% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigenetic-funcs.R
\name{epigen_add}
\alias{epigen_add}
\title{Add epigenetic data to MOAS like data}
\usage{
epigen_add(MOAS, file_path,
  match_path = "~/LCBC/Projects/Cross_projects/MOAS/data-raw/DNA/gID_MOAS_match.tsv",
  debug = FALSE)
}
\arguments{
\item{MOAS}{A MOAS-derived data frame you want data added to}

\item{file_path}{path to excel file containing epigenetic
data}

\item{match_path}{path to file that matches genetic ID to
CrossProject_ID}

\item{debug}{logical, if debugging informaition should be added.
Defaults to FALSE}
}
\description{
Add epigenetic data to MOAS like data
}
\seealso{
Other epigen-functions: \code{\link{epigen_get}}

Other MOAS add-functions: \code{\link{pgs_add_all}},
  \code{\link{pgs_add}}, \code{\link{pgs_get_all}}
}
\concept{MOAS add-functions}
\concept{epigen-functions}
