% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_times.R
\name{factor_times}
\alias{factor_times}
\title{Factor time of day into categories of 4.}
\usage{
factor_times(TimeVector)
}
\arguments{
\item{TimeVector}{String or time data in 'HH:MM' format}
}
\value{
A string vector with up to 4 levels.
}
\description{
\code{factor_times} will categorise a time vector into one of
four times of day: Morning, Afternoon, Evening, or Night.
}
\examples{
 factor_times(data)

}
