% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_sensitive}
\alias{filter_sensitive}
\title{Filter sensitive data out of the MOAS}
\usage{
filter_sensitive(data)
}
\arguments{
\item{data}{The MOAS or a MOAS generated file.}
}
\value{
tibble
}
\description{
Filter sensitive data out of the MOAS
}
