% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_site}
\alias{filter_site}
\title{Utility function for reducing double/triple scans to a single row}
\usage{
filter_site(
  data,
  site_var = site_name,
  keep = "long",
  tie = "interval",
  site_order = c("ousPrisma", "ousSkyra", "ousAvanto"),
  verbose = TRUE
)
}
\arguments{
\item{data}{data extracted from the NOAS}

\item{site_var}{column name with the site information}

\item{keep}{A string specifying which data from double/triple scans to keep.}

\item{tie}{string indicating given a tie in the "long" keep option, what to keep.}

\item{site_order}{string vector of the scanner priority given a tie between scanners}

\item{verbose}{logical, TRUE sets it to verbose}
}
\value{
A MOAS type file with one line per subject and timepoint.
}
\description{
returns a MOAS data.frame with one row per
participant and timepoint (i.e. removes double/triple scan entries).
For analyses not intending to use the power of double/triple scans,
or for data which will be widened and scanner type/site is not of
interest.
Only subject timepoints that have several entries
(i.e. several scan sites),
will be reduced. All timepoints will be retained.
}
\details{
Available options for 'keep' are:
\itemize{
 \item{"long"}{keep data from scanner with most data (default)}
 \item{"ousAvanto"}{keep 'ousAvanto'}
 \item{"ousSkyra"}{keep 'ousSkyra'}
 \item{"ousPrisma"}{'ousPrisma'}
}

Available options for 'tie' are:
\itemize{
 \item{"interval"}{keep data from scanner with longest data interval (default)}
 \item{"ousAvanto"}{keep 'ousAvanto'}
 \item{"ousSkyra"}{keep 'ousSkyra'}
 \item{"ousPrisma"}{'ousPrisma'}
}
}
\examples{
\dontrun{
filter_site(noas_data)
filter_site(noas_data, 'ousSkyra')
filter_site(noas_data, 'ousAvanto')
}
}
