% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_trainingexposed}
\alias{filter_trainingexposed}
\title{Filter out when a participant has trained}
\usage{
filter_trainingexposed(data, predicate)
}
\arguments{
\item{data}{The MOAS or a MOAS generated file.}

\item{predicate}{a logical statement to identify
rows of data under memory experimentation}
}
\value{
tibble
}
\description{
Several LCBC projects have some memory training
experimental procedure. For some papers, this
might come in the way of answering hypotheses.
This function helps filter out participants data
after they have been exposed to training. Meaning
data before training remain, data after do not.
}
\examples{
  dt <-  data.frame(
    CrossProject_ID = rep("1000000", 6),
    Site_Name = c("ousAvanto", "ousAvanto","ousAvanto",
                "ousSkyra", "ousSkyra", "ousSkyra"),
    Subject_Timepoint = c(1:3,3:5),
    Age = c(8, 10, 14, 14, 17, 20),
    stringsAsFactors = FALSE
  )
  filter_site(dt)
}
