% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_times.R
\name{fix_times}
\alias{fix_times}
\title{Calculates time in decimal hours from HH:MM format.}
\usage{
fix_times(data)
}
\arguments{
\item{data}{String or time data in 'HH:MM' format}
}
\value{
a numeric vector of decimal hours
}
\description{
\code{fix_times} calculates time in decimal hours from HH:MM format.
}
\examples{
 fix_times(data)

}
