% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_lmm.R
\name{fs_lmm}
\alias{fs_lmm}
\title{Create a data.frame that is ready for use in Freesufers linear
mixed models}
\usage{
fs_lmm(data, formula, site_var, folder_var, missing = "mean", ..., file = NULL)
}
\arguments{
\item{data}{data extracted from the NOAS}

\item{formula}{right-hand formula for your model. (ex. ~ age * sex)}

\item{site_var}{column name with the site information}

\item{folder_var}{unqoted column name with folder-information of MRI data
from the 'mri_info' table.}

\item{missing}{Action to take on missing data that are numeric}

\item{...}{other arguments to \code{\link{utils}{model.matrix}}}

\item{file}{Optional string specifying file name to be saved}
}
\value{
a data frame ready for Freesurfer LMM use.
}
\description{
\code{fs_lmm} Creates a data.frame of subsetted row and formatted
columns ready for use in Freesurfers (6.0) linear mixed models.
Design matrices are created for the grouping.var factors, and numeric
variables are z-transformed.
}
\details{
Available options for 'missing.action' are:
\itemize{
 \item{"delete"}{delete observations with any missing numeric.vars}
 \item{"mean"}{replace missing with mean for that participant}
 \item{"all"}{replace all values with the mean for that participant}
 \item{"first"}{replace all values with the first
 observation for that participant}
}
}
\examples{
\dontrun{
fs_lmm(data,
       age * sex * cog
     )
}
}
