% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic-funcs.R
\name{get_pgs_all}
\alias{get_pgs_all}
\title{Get all PGS's}
\usage{
get_pgs_all(s_levels = paste0("S", 1:12),
  pgs_path = "~/LCBC/Projects/Cross_projects/Genetics/PGS/PGS_20190618/PGS_wAPOE/",
  genetic_match_file = "~/LCBC/Projects/Cross_projects/MOAS/data-raw/DNA/gID_MOAS_match.tsv",
  include_cnt = FALSE, include_genetics_debug = FALSE)
}
\arguments{
\item{s_levels}{character vector of the significance levels wanted}

\item{pgs_path}{path to the directory where the PGS's are stored}

\item{genetic_match_file}{path to the file containing the MOAS-genetics
matching and debugging information}

\item{include_cnt}{logical, whether to keep SNP count information}

\item{include_genetics_debug}{logical, whether to keep all columns
in the genetic_match_file in the final output}
}
\value{
a tibble / data.frame
}
\description{
This function calls on [\code{get_pgs}] to read in wanted
all PGS at specified significance levels (by default all 12)
}
\examples{
\dontrun{

# Here we assume you know the path to the LCBC
# lagringshotell, you can substitute "~" in the
# paths with the path to the lagringshotell

# In this version you dont need to specify the
# PGS you want, it will take all it finds in the
# folder path provided.
#

get_pgs_all(
    s_levels = c("S1", "S7", "S11"),
    pgs_path = "~/LCBC/Projects/Cross_projects/Genetics/PGS/PGS_20190618/PGS_wAPOE/",
    genetic_match_file = "~/LCBC/Projects/Cross_projects/MOAS/data-raw/DNA/gID_MOAS_match.tsv",
)
}
}
