% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggaparc.R
\name{ggaparc}
\alias{ggaparc}
\title{Plot Freesurfer aparc areas}
\usage{
ggaparc(data = NULL, plot.areas = NULL, mapping = NULL,
  colour = "white", size = 0.3, show.legend = NA, na.fill = "grey",
  ...)
}
\arguments{
\item{data}{A data.frame to use for plot aesthetics. Must include a
column called "area" corresponding to aparc areas.}

\item{plot.areas}{Character vector, plots only areas specified in the vector.}

\item{mapping}{ggplot2 aethetics (cannot include x and y aethetics)}

\item{colour}{string or HEX code for the colour of the outlines of each area}

\item{size}{Numeric, size of the line outlining each area}

\item{show.legend}{logical, toggle on or off legend.}

\item{na.fill}{string or HEX code for the fill of the area without values}

\item{...}{}
}
\value{
a ggplot object
}
\description{
\code{ggaparc} plots and returns a ggplot object of plotted
aparc areas.
}
\examples{
ggaparc()
ggaparc(na.fill = "transparent",mapping=aes(fill=area))
ggaparc(colour="black", size=.7, mapping=aes(fill=area)) + theme_void()
}
\seealso{
[ggplot()], [aes()], [geom_polygon()], [coord_fixed()] from the ggplot2 package
}
\author{
Athanasia Mowinckel and Didac Pineiro
}
