% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbody-funcs.R
\name{inbody_add}
\alias{inbody_add}
\title{Add InBody data to data}
\usage{
inbody_add(data, path, suffix = ".x", ...)
}
\arguments{
\item{data}{MOAS-like data}

\item{path}{path to file}

\item{suffix}{suffix to add if colum names conflict}

\item{...}{Additional arguments to [/code{inbody_get}]}
}
\value{
data frame
}
\description{
Reads and cleans InBody data from a
file and adds it to MOAS like data.
}
\examples{
\dontrun{

path <- "~/MOAS/data-raw/InBody/OrigData/2019_10/inbody770_data.csv"
inbody_add(MOAS, path)
inbody_add(MOAS, path, keep.limits = TRUE)

MOAS \%>\%
   inbody_add(path, keep.kHz= TRUE)
}
}
