% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbody-funcs.R
\name{inbody_add_all}
\alias{inbody_add_all}
\title{Add all InBody data from folder to data}
\usage{
inbody_add_all(data, path, suffix = ".x", ...)
}
\arguments{
\item{data}{MOAS-like data}

\item{path}{folder path}

\item{suffix}{suffix to add if colum names conflict}

\item{...}{additional arguments to [/code{inbody_get}]}
}
\value{
data frame
}
\description{
Add all InBody data contained in a folder
to MOAS-like data provided
}
\examples{
\dontrun{

path <- "~/MOAS/data-raw/InBody/OrigData/"
inbody_add_all(MOAS, path)
inbody_add_all(MOAS, path, keep.limits = TRUE)

MOAS \%>\%
   inbody_add_all(path, keep.kHz= TRUE)
}
}
