% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbody-funcs.R
\name{inbody_get}
\alias{inbody_get}
\title{Get InBody data}
\usage{
inbody_get(
  path,
  keep.limits = FALSE,
  keep.kHz = FALSE,
  keep.partials = FALSE,
  keep.indices = FALSE
)
}
\arguments{
\item{path}{path to file}

\item{keep.limits}{logical - Keep "Limit" columns or not}

\item{keep.kHz}{logical - Keep "kHz" columns or not}

\item{keep.partials}{logical - Keep "of" columns or not}

\item{keep.indices}{logical - Keep columns Levels/Scores etc. or not}
}
\value{
data frame
}
\description{
Will read, and do some initial cleaning of
the InBody data. Calls two other internal functions
to read the InBody data that are either pre-cleaned
or completely raw. Will choose for you.
}
\examples{
\dontrun{

path <- "~/MOAS/data-raw/InBody/OrigData/2019_10/inbody770_data.csv"
inbody_get(path)
inbody_get(path, keep.limits = TRUE)
}
}
