% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbody-funcs.R
\name{inbody_get_all}
\alias{inbody_get_all}
\title{Get all Inbody data from folder}
\usage{
inbody_get_all(path, ...)
}
\arguments{
\item{path}{folder path}

\item{...}{additional arguments to [/code{inbody_get}]}
}
\value{
data frame
}
\description{
Will given a folder directory,
read, clean and return all data contained
in inbody770.csv files.
}
\examples{
\dontrun{

path <- "~/MOAS/data-raw/InBody/OrigData/"
inbody_get_all(path)
inbody_get_all(path, keep.limits = TRUE)
}
}
