% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbody-funcs.R
\name{inbody_read}
\alias{inbody_read}
\title{Read InBody data}
\usage{
inbody_read(path)
}
\arguments{
\item{path}{path to file}
}
\value{
data frame
}
\description{
This function will read in any InBody
file, whether it has been cleaned manually, or
is completely raw and unaltered. Since this
is intended for the MOAS, is assumes IDs are stored
in a meaningful way in accordance with LCBC
data collection principles <ID.proj.wave>
}
\examples{
\dontrun{

path <- "~/MOAS/data-raw/InBody/OrigData/2019_10/inbody770_data.csv"
inbody_read(path)

}
}
