% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.hms.R
\name{is.hms}
\alias{is.hms}
\title{Utility function to locate hms columns}
\usage{
is.hms(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
a data frame with no empty columns
}
\description{
\code{is.hms} Removes columns from the
data frame that does not contain any data.
}
\examples{
is.hms(data)

}
