% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-funcs.R
\name{is_hms}
\alias{is_hms}
\title{Utility function to locate hms columns}
\usage{
is_hms(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
logical vector of length==ncol(data)
}
\description{
\code{is_hms} locates columns that are time (hms) classes
}
\examples{
\dontrun{
is_hms(data)
}
}
