% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tp.R
\name{launch_check_tp}
\alias{launch_check_tp}
\title{Launch shiny app to find subject timepoint}
\usage{
launch_check_tp(...)
}
\arguments{
\item{...}{arguments to [\code{shiny::runApp}]}
}
\description{
A companion function to [\code{check_tp}], which
launches an interactive shiny app to find participant ID.
}
