% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_LCBCshiny.R
\name{launch_LCBCshiny}
\alias{launch_LCBCshiny}
\title{Function to launch the LCBC shiny data explorer}
\usage{
launch_LCBCshiny(data, ...)
}
\arguments{
\item{data}{The MOAS or a MOAS generated file.}

\item{...}{any arguments to [\code{shiny::runApp}]}
}
\description{
\code{launch_LCBCshiny} will launch the shiny app made to explore
and extract data from the MOAS.
Running it without supplied data will enable you to upload the
data you wish to explore via the browser.
If you have the data already loaded in your R environment, you
may pass this as an argument to the function to have the data
already loaded in the application.
}
\examples{
\dontrun{
launch_LCBCshiny()

launch_LCBCshiny(MOAS)
}
}
