% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-lcbc.R
\name{lcbc_cols}
\alias{lcbc_cols}
\title{Function to extract lcbc colors as hex codes}
\usage{
lcbc_cols(...)
}
\arguments{
\item{...}{Character names of lcbc_colors}
}
\description{
Function to extract lcbc colors as hex codes
}
