% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-logo.R
\name{lcbc_logo}
\alias{lcbc_logo}
\title{Add LCBC logo to plot}
\usage{
lcbc_logo(type = "main", alpha = 0.4, ...)
}
\arguments{
\item{type}{type of logo to get}

\item{alpha}{opacity of logo}

\item{...}{arguments to \link{ggplot2}{\code{annotation_custom}}}
}
\description{
in a ggplot, adding the logo should be added
early in the plot building
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = disp, colour = cyl)) +
  geom_point() +
  lcbc_logo()
}
