% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-lcbc.R
\name{lcbc_pal}
\alias{lcbc_pal}
\title{Return function to interpolate a lcbc color palette}
\usage{
lcbc_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in lcbc_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a lcbc color palette
}
