% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_date}
\alias{mean_date}
\title{Calculate the mean between two dates}
\usage{
mean_date(date1, date2)
}
\arguments{
\item{date1}{first date}

\item{date2}{second date}
}
\value{
mean difference between two dates
}
\description{
Calculate the mean between two dates
}
\examples{
dates <- as.Date(c("2020-01-20", "1998-01-20"))
mean_date(dates[1], dates[2])
}
