% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutates.R
\name{mutate_ages}
\alias{mutate_ages}
\title{Calculates ages based on birthdates and test/MRI dates}
\usage{
mutate_ages(data)
}
\arguments{
\item{data}{The MOAS or a MOAS generated file.}
}
\value{
a data frame with added/replaced age and timepoint variables.
}
\description{
\code{mutate_ages} Creates a dataframe where age and subject timepoint
have been added to the data. Both MRI and Test age are calculated,
and 'Age' is the mean between these two. If only one of MRI or Test
age are present, this one is used for Age.
}
\examples{
\dontrun{
mutate_ages(data)
}

}
