% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.col.rm.R
\name{na.col.rm}
\alias{na.col.rm}
\title{Utility function to remove columns with no data}
\usage{
na.col.rm(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
a data frame with no empty columns
}
\description{
\code{na.col.rm} Removes columns from the
data frame that does not contain any data.
}
\examples{
\dontrun{
na.col.rm(data)
}
}
