% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moas-utils.R
\name{na_col_rm}
\alias{na_col_rm}
\title{Utility function to remove columns with no data}
\usage{
na_col_rm(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
a data frame with no empty columns
}
\description{
\code{na.col.rm} Removes columns from the
data frame that does not contain any data.
}
\examples{
df <- data.frame(
ID = 1:3,
Age = c(22, 25,17),
Sex = NA,
group = NA
)
na_col_rm(df)
}
