% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noas_example}
\alias{noas_example}
\title{Example NOAS data}
\format{
A data frame with 10 rows and 8 variables:
\describe{
  \item{subject_id}{subject identifier}
  \item{project_id}{project identifier}
  \item{wave_code}{assessment wave for the project}
  \item{site_name}{mri site identifier}
  \item{mri_info_folder}{mri folder on TSD}
  \item{visit_age}{subject age at visit}
  \item{cog}{cognitive score}
  \item{sex}{subject sex}
}
}
\usage{
noas_example
}
\description{
A dataset containing mock data mimicking
core features like the NOAS data. For testing
and examples in the package.
}
\keyword{datasets}
