% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgs-funcs.R
\name{pgs_add_all}
\alias{pgs_add_all}
\title{Add all PGS data to the MOAS}
\usage{
pgs_add_all(MOAS = NULL, s_levels = paste0("S", 1:12),
  pgs_path = NULL, genetic_match_file = NULL, include_cnt = FALSE,
  include_genetic_debug = FALSE)
}
\arguments{
\item{MOAS}{data.frame of the full MOAS or MOAS sub-set}

\item{s_levels}{character vector of the significance levels wanted}

\item{pgs_path}{path to the directory where the PGS's are stored}

\item{genetic_match_file}{path to the file containing the MOAS-genetics
matching and debugging information}

\item{include_cnt}{logical, whether to keep SNP count information}

\item{include_genetic_debug}{logical, whether to keep all columns
in the genetic_match_file in the final output}
}
\value{
a tibble / data.frame
}
\description{
This function calls on [\code{pgs_add}] to read in all
PGS at specified significance levels, and adds that data
to the MOAS-type data provided.
}
\examples{
\dontrun{

# Here we assume you know the path to the LCBC
# lagringshotell, you can substitute "~" in the
# paths with the path to the lagringshotell

# In this version you dont need to specify the
# PGS you want, it will take all it finds in the
# folder path provided, and add them to the MOAS
# data you provided

pgs_add_all(
    MOAS,
    s_levels = c("S1", "S7", "S11"),
    pgs_path = "path/to/PGS_wAPOE/",
    genetic_match_file = character(),
)
}
}
\seealso{
Other pgs-functions: \code{\link{pgs_add_single}},
  \code{\link{pgs_add}}, \code{\link{pgs_get_all}},
  \code{\link{pgs_get_single}}, \code{\link{pgs_get}},
  \code{\link{pgs_read}}

Other MOAS add-functions: \code{\link{epigen_add}},
  \code{\link{pgs_add_single}}, \code{\link{pgs_add}},
  \code{\link{pgs_get_all}}
}
\concept{MOAS add-functions}
\concept{pgs-functions}
