% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgs-funcs.R
\name{pgs_add_single}
\alias{pgs_add_single}
\title{Add single PGS data to the MOAS}
\usage{
pgs_add_single(MOAS = NULL, pgs_file = character(),
  genetic_match_file = character(), include_cnt = FALSE,
  include_genetic_debug = FALSE)
}
\arguments{
\item{MOAS}{data.frame of the full MOAS or MOAS sub-set}

\item{pgs_file}{path to PGS.profile}

\item{genetic_match_file}{path to the file containing the MOAS-genetics
matching and debugging information}

\item{include_cnt}{logical, whether to keep SNP count information}

\item{include_genetic_debug}{logical, whether to keep all columns
in the genetic_match_file in the final output}
}
\value{
a tibble / data.frame
}
\description{
This function calls on [\code{pgs_add}] to read in all
PGS at specified significance levels, and adds that data
to the MOAS-type data provided.
}
\examples{
\dontrun{

# In this version you dont need to specify the
# PGS.profile you want

pgs_add_single(
    MOAS
    pgs_file = "path/to/Genetics/PGS/PGS_somedate/",
    genetic_match_file = "path/to//gID_MOAS_match.tsv",
)
}
}
\seealso{
Other pgs-functions: \code{\link{pgs_add_all}},
  \code{\link{pgs_add}}, \code{\link{pgs_get_all}},
  \code{\link{pgs_get_single}}, \code{\link{pgs_get}},
  \code{\link{pgs_read}}
}
\concept{pgs-functions}
