% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgs-funcs.R
\name{pgs_get_all}
\alias{pgs_get_all}
\title{Get all PGS's}
\usage{
pgs_get_all(s_levels = paste0("S", 1:12), pgs_path = NULL,
  genetic_match_file = NULL, include_cnt = FALSE,
  include_genetic_debug = FALSE)
}
\arguments{
\item{s_levels}{character vector of the significance levels wanted}

\item{pgs_path}{path to the directory where the PGS's are stored}

\item{genetic_match_file}{path to the file containing the MOAS-genetics
matching and debugging information}

\item{include_cnt}{logical, whether to keep SNP count information}

\item{include_genetic_debug}{logical, whether to keep all columns
in the genetic_match_file in the final output}
}
\value{
a tibble / data.frame
}
\description{
This function calls on [\code{pgs_get}] to read in wanted
all PGS at specified significance levels (by default all 12)
}
\examples{
\dontrun{

# Here we assume you know the path to the LCBC
# lagringshotell, you can substitute "~" in the
# paths with the path to the lagringshotell

# In this version you dont need to specify the
# PGS you want, it will take all it finds in the
# folder path provided.
#

pgs_get_all(
    s_levels = c("S1", "S7", "S11"),
    pgs_path = "path/to/Genetics/PGS/PGS_20190618/PGS_wAPOE/",
    genetic_match_file = "path/to/MOAS/data-raw/DNA/gID_MOAS_match.tsv",
)
}
}
\seealso{
Other pgs-functions: \code{\link{pgs_add_all}},
  \code{\link{pgs_add_single}}, \code{\link{pgs_add}},
  \code{\link{pgs_get_single}}, \code{\link{pgs_get}},
  \code{\link{pgs_read}}

Other MOAS add-functions: \code{\link{epigen_add}},
  \code{\link{pgs_add_all}}, \code{\link{pgs_add_single}},
  \code{\link{pgs_add}}
}
\concept{MOAS add-functions}
\concept{pgs-functions}
