% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgs-funcs.R
\name{pgs_get_single}
\alias{pgs_get_single}
\title{Read in and sort a single PGS profile}
\usage{
pgs_get_single(pgs_file = character(),
  genetic_match_file = character(), include_cnt = FALSE,
  include_pheno = FALSE, include_genetic_debug = FALSE)
}
\arguments{
\item{pgs_file}{path to PGS.profile}

\item{genetic_match_file}{path to the file containing the MOAS-genetics
matching and debugging information}

\item{include_cnt}{logical, whether to keep SNP count information}

\item{include_pheno}{logical, whether to keep PHENO information}

\item{include_genetic_debug}{logical, whether to keep all columns
in the genetic_match_file in the final output}
}
\value{
a tibble / data.frame
}
\description{
Given the path containing PGS profile, this function will read in
PGS data for the singe pgs file. In order to correctly only keep
rows of data from verified sources, the genetic_match_file is
necessary to provide for a check.
}
\examples{
\dontrun{
pgs_get_single( pgs_path = "path/to//PGS_wAPOE/",
         genetic_match_file =  "path/to/gID_MOAS_match.tsv",
)

# You can also toggle adding the CNT columns from the PGS, by changing
# include_cnt to TRUE
pgs_get_single( pgs_path = "path/to/PGS_wAPOE/",
         genetic_match_file = "path/to/gID_MOAS_match.tsv",
         include_cnt = TRUE
)

}

}
\seealso{
Other pgs-functions: \code{\link{pgs_add_all}},
  \code{\link{pgs_add_single}}, \code{\link{pgs_add}},
  \code{\link{pgs_get_all}}, \code{\link{pgs_get}},
  \code{\link{pgs_read}}

Other MOAS get-functions: \code{\link{dbs_get}},
  \code{\link{epigen_get}}, \code{\link{pgs_get}}
}
\concept{MOAS get-functions}
\concept{pgs-functions}
