% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgs-funcs.R
\name{pgs_read}
\alias{pgs_read}
\title{Read in a PGS file}
\usage{
pgs_read(path, name = NULL)
}
\arguments{
\item{path}{path to PGS file}

\item{name}{Name of the PGS}
}
\value{
data frame
}
\description{
Read in a PGS file
}
\seealso{
Other pgs-functions: 
\code{\link{pgs_add_all}()},
\code{\link{pgs_add_single}()},
\code{\link{pgs_add}()},
\code{\link{pgs_get_all}()},
\code{\link{pgs_get_single}()},
\code{\link{pgs_get}()}
}
\concept{pgs-functions}
