% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-projects.R
\name{project_cols}
\alias{project_cols}
\title{Function to extract project colors as hex codes}
\usage{
project_cols(...)
}
\arguments{
\item{...}{Character names of project_colors}
}
\description{
Function to extract project colors as hex codes
}
