% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Projects}
\alias{Projects}
\title{MOAS Projects}
\format{An object of class \code{data.frame} with 9 rows and 3 columns.}
\usage{
data(Projects)
}
\description{
data frame of LCBC projects in the MOAS
}
\examples{
data(Projects)
}
\keyword{datasets}
