% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scales.R
\name{scale_color_lcbc}
\alias{scale_color_lcbc}
\title{Color scale constructor for lcbc colors}
\usage{
scale_color_lcbc(...)
}
\arguments{
\item{...}{arguments to \code{\link{scale_colour_lcbc}}}
}
\description{
Color scale constructor for lcbc colors
}
