% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scales.R
\name{scale_color_proj}
\alias{scale_color_proj}
\title{Color scale constructor for proj colors}
\usage{
scale_color_proj(...)
}
\arguments{
\item{...}{arguments to \code{\link{scale_colour_proj}}}
}
\description{
Color scale constructor for proj colors
}
