% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scales.R
\name{scale_fill_lcbc}
\alias{scale_fill_lcbc}
\title{Fill scale constructor for lcbc colours}
\usage{
scale_fill_lcbc(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in lcbc_palettes}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to \code{\link{ggplot2}{discrete_scale()}} or
\code{\link{ggplot2}{scale_colour_gradientn()}}, used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for lcbc colours
}
