% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{select_sensitive}
\alias{select_sensitive}
\title{Select away sensitive columns}
\usage{
select_sensitive(data, remove = TRUE)
}
\arguments{
\item{data}{The MOAS or a MOAS generated file.}

\item{remove}{logical, to remove (default) sensitive columns,
or select them}
}
\value{
tibble
}
\description{
This function will remove all the columns from
a MOAS-like data.frame that is known to include
possible sensitive information.
}
