% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sites}
\alias{Sites}
\title{MOAS sites}
\format{
An object of class \code{data.frame} with 5 rows and 3 columns.
}
\usage{
data(Sites)
}
\description{
a data frame with information on LCBC sites
}
\examples{
data(Sites)
}
\keyword{datasets}
