% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\name{theme_lcbc_dark_void}
\alias{theme_lcbc_dark_void}
\title{ggplot minimal theme for lcbc}
\usage{
theme_lcbc_dark_void(base_size = base_size_text)
}
\arguments{
\item{base_size}{text size}
}
\description{
ggplot minimal theme for lcbc
}
