% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\name{theme_lcbc_grid}
\alias{theme_lcbc_grid}
\title{ggplot theme for lcbc}
\usage{
theme_lcbc_grid(base_size = base_size_text)
}
\arguments{
\item{base_size}{text size}
}
\description{
ggplot theme for lcbc
}
