% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdify.R
\name{tsdify}
\alias{tsdify}
\title{Creates TSD folder names for MRI data location, or
project wave from TSD folder names}
\usage{
tsdify(DATA, reverse = F)
}
\arguments{
\item{reverse}{logical. If FALSE (default) creates TSD folder
strings. If TRUE, finds project wave number from Folder information.}

\item{data}{The MOAS or a MOAS generated file.}
}
\value{
either an string vector of folder names (default), or a
vector of integers
}
\description{
\code{tsdify} is a function to use to either figure out what the
TSD folder location of MRI data is, or to figure out the project
wave from TSD folder names.
}
\examples{
 tsdify(Data)
 tsdify(Data, reverse=TRUE)

}
