% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variables}
\alias{variables}
\title{MOAS variables}
\format{
An object of class \code{data.frame} with 835 rows and 5 columns.
}
\usage{
data(variables)
}
\description{
a data frame with information on the different MOAS variables and
their content and types
}
\examples{
data(variables)
}
\keyword{datasets}
