% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widen.R
\name{widen}
\alias{widen}
\title{Widen the MOAS data.frame}
\usage{
widen(data, by, keep = NULL)
}
\arguments{
\item{data}{The MOAS or a MOAS generated file.}

\item{by}{Column which to widen by}

\item{keep}{Option to provide to [site_keeper()] for
filtering the double/triple scanned}
}
\value{
A MOAS type file widened by selected column,
prefixed with columns specifications
}
\description{
This is a function to create a widened data.frame from the MOAS.
Widening of the data is necessary for performing ANOVAs and
other statistical analysis, particularly if intending to run
in SPSS.
}
\details{
For MOAS which inclused double/triple scans, by="Site_Name",
must be run prior to by="Project_Wave".

'by' options are:
\itemize{
\item{Subject_Timepoint}
\item{Project_Wave}
\item{Site_Name}
\item{Site_Number}
}
}
\examples{
\dontrun{
widen(MOAS, by="Site_Name")
widen(MOAS, by="Project_Wave")

# Data with double/triple scans must be widened
# in a two-stage process
dt = widen(MOAS, by="Site_Name")
df = widen(dt, by="Project_Wave")
}

}
