% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yeo7.R
\docType{data}
\name{Yeo7}
\alias{Yeo7}
\title{Yeo 7 Resting-state Cortical Parcellations}
\format{A data.frame with columns for area, hemisphere (left, right) and view (lateral, medial)}
\usage{
data(Yeo7)
}
\description{
Coordinate data for the resting-state networks of
the Yeo 2011 7 networks.
}
\examples{
data(Yeo7)
}
\references{
Yeo et al. (2011) J. Neurophysiology 16(3):1125-1165
(\href{https://www.ncbi.nlm.nih.gov/pubmed/21653723}{PubMed})

\describe{
  \item{lat}{coordinates for the x-axis}
  \item{long}{coordinates for the y-axis}
  \item{area}{description of Col1}
  \item{hemi}{name of the hemisphere (left, right)}
  \item{side}{which side to view (medial, lateral)}
}
}
\keyword{datasets}
