% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_timepoint}
\alias{add_timepoint}
\title{Add timepoint to data}
\usage{
add_timepoint(data, name = timepoint)
}
\arguments{
\item{data}{data extracted from the NOAS}

\item{name}{unquoted name to give the new column}
}
\value{
data frame with extra column with timepoint
}
\description{
Calculated sequential timepoint
for participants along the 'age'
column.
}
\examples{
# attach built-in noas example data to test
dt <- noas_example

add_timepoint(dt)

library(dplyr)
dt \%>\%
 add_timepoint()

# Change the name of the variable
dt \%>\%
 add_timepoint(name = tp)

}
