% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-funcs.R
\name{factor_times}
\alias{factor_times}
\title{Factor time of day}
\usage{
factor_times(x)
}
\arguments{
\item{x}{character or hms vector}
}
\value{
factor vector
}
\description{
Takes a vector of HH:MM (HH:MM:SS) information and
categorises these by a 4 level factor of time of day.
}
\examples{
factor_times(c("12:23", "15:59", "22:10", "8:13"))
}
