% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_lmm.R
\name{fs_lmm}
\alias{fs_lmm}
\title{Create a data.frame that is ready for use in Freesufers linear
mixed models}
\usage{
fs_lmm(data, grouping.var = "Site_Name", numeric.var = "Age",
  missing.missing.action = "Mean", keep = "long", file = NULL)
}
\arguments{
\item{data}{The MOAS or a MOAS generated file.}

\item{grouping.var}{String vector specifying the column names of
categorical/ordinal factors.}

\item{numeric.var}{A string vector of column names for numeric,
scalar co-variates.}

\item{keep}{For double/triple scans, which data should be kept.
option calls \code{site_keeper}.}

\item{file}{Optional string specifying file name to be saved,
omit extention.}

\item{missing.action}{Action to take on missing data for the
numeric.vars. Options are:
'delete' - deletes observations with any missing numeric.vars,
'Mean' - replaces missing with mean for that participant,
'All' - replaces all values with the mean for that participant,
'first' - replaced all values with the first observation for that
participant.}
}
\value{
a data frame ready for Freesurfer LMM use.
}
\description{
\code{fs_lmm} Creates a data.frame of subsetted row and formatted
columns ready for use in Freesurfers (6.0) linear mixed models.
Design matrices are created for the grouping.var factors, and numeric
variables are z-transformed.
}
\examples{
site_keeper(MOAS)
site_keeper(MOAS, "ousSkyra")
site_keeper(MOAS, "ousAvanto")

}
