% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_lmm.R
\name{fs_lmm}
\alias{fs_lmm}
\title{Create a data.frame that is ready for use in Freesufers linear
mixed models}
\usage{
fs_lmm(
  data,
  formula,
  site_var,
  folder_var,
  numeric_transform = "delete",
  ...,
  file = NULL,
  concat_list = NULL
)
}
\arguments{
\item{data}{data extracted from the NOAS}

\item{formula}{right-hand formula for your model. (ex. ~ age * sex)}

\item{site_var}{column name with the site information}

\item{folder_var}{unqoted column name with folder-information of MRI data
from the 'mri_info' table.}

\item{numeric_transform}{Action to take on numeric_transform data that are numeric}

\item{...}{other arguments to \code{\link{utils}{model.matrix}}}

\item{file}{Optional string specifying file name to be saved}

\item{concat_list}{character vector of fsid's that you want the data matched to.
Used if the data is already concatenated and you wish to add more variables to
your models.}
}
\value{
a data frame ready for Freesurfer LMM use.
}
\description{
\code{fs_lmm} Creates a data.frame of subsetted row and formatted
columns ready for use in Freesurfers (6.0) linear mixed models.
Design matrices are created for the grouping.var factors, and numeric
variables are z-transformed.
}
\details{
Available options for 'numeric_transform' are:
\itemize{
 \item{"delete"}{delete observations with any numeric_transform numeric.vars}
 \item{"mean_na"}{replace numeric_transform with mean for that participant}
 \item{"mean_all"}{replace all values with the mean for that participant}
 \item{"first"}{replace all values with the first
 observation for that participant}
}
}
\examples{
# attach built-in noas example data to test
dt <- noas_example

fs_lmm(dt, ~ age * sex * cog,
       site_var = site_name,
       folder_var = folder
     )

 # replace NA values in numeric with
 # mean values for the participant
 fs_lmm(dt, ~ age * sex * cog,
       numeric_transform = "mean_na",
       site_var = site_name,
       folder_var = folder
     )

 # replace all numeric values with
 # mean values for the participant
 fs_lmm(dt, ~ age * sex * cog,
       numeric_transform = "mean_na",
       site_var = site_name,
       folder_var = folder
     )

 # replace all numeric values with
 # first for the participant
 fs_lmm(dt, ~ age * cog,
       site_var = site_name,
       folder_var = folder
     )

 # Provide a vector of fsid to reduce the data to
 # pre-existing concatenated imaging data.
 fs_lmm(noas_example, ~ age,
        site_var = site_name,
        folder_var = folder,
        concat_list = c("1000000_1", "1000000_3", "1000000_5")
        )

}
