% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_brain.R
\name{theme_brain}
\alias{theme_brain}
\alias{theme_darkbrain}
\title{ggbrain plot theme}
\usage{
theme_brain()

theme_darkbrain()
}
\description{
a set of themes created for the ggbrain plots. Use ggplot2::theme() to tweak.
}
\details{
\describe{

\item{`theme_brain`}{
Default theme for ggbrain. Transparent background, no axis lines, and no grid.}

\item{`theme_darkbrain`}{
Dark equivalent to theme_brain, with black background, and light text.}
}
}
\seealso{
[ggplot()], [aes()], [geom_polygon()], [coord_fixed()] from the ggplot2 package
}
\author{
Athanasia Mowinckel
}
