% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.hms.R
\name{is.hms}
\alias{is.hms}
\title{Utility function to locate hms columns}
\usage{
is.hms(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
logical vector of length==ncol(data)
}
\description{
\code{is.hms} locates columns that are time (hms) classes
}
\examples{
\dontrun{
is.hms(data)
}
}
