% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{mutate_sensitive}
\alias{mutate_sensitive}
\title{Alter sensitive columns to make them less sensitive}
\usage{
mutate_sensitive(data, scramble_ids = TRUE)
}
\arguments{
\item{data}{The MOAS or a MOAS generated file.}

\item{scramble_ids}{make IDs anonymous}
}
\value{
tibble
}
\description{
THe function will round Age to nearest integer, and
scramble participants ids, to make data less sensitive
}
