% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R, R/site_keeper.R
\name{site_keeper}
\alias{site_keeper}
\title{Deprecated, use filter_site}
\usage{
site_keeper(data, keep = "long", tie = "interval",
  site_order = c("ousPrisma", "ousSkyra", "ousAvanto"), quiet = F)

site_keeper(data, keep = "long", tie = "interval",
  site_order = c("ousPrisma", "ousSkyra", "ousAvanto"), quiet = F)
}
\arguments{
\item{data}{The MOAS or a MOAS generated file.}

\item{keep}{A string specifying which data from double/triple scans to keep.
Available options are:
'long' - keep data from scanner with most data (default),
'ousAvanto' = keep 'ousAvanto',
'ousSkyra' = keep 'ousSkyra', or
'ousPrisma' = keep 'ousPrisma'.}

\item{tie}{string indicating given a tie in the "long" keep option, what to keep.
Available options are:
'interval' - keep data from scanner with longest data interval (default),
'ousAvanto' = keep 'ousAvanto',
'ousSkyra' = keep 'ousSkyra', or
'ousPrisma' = keep 'ousPrisma'.}

\item{site_order}{string vector of the scanner priority given a tie between scanners}

\item{quiet}{logical, TRUE sets it to verbose}

\item{...}{any argument to filter_site}
}
\value{
A MOAS type file with one line per subject and timepoint.
}
\description{
\code{site_keeper} returns a MOAS data.frame with one row per
participant and timepoint (i.e. removes double/triple scan entries).
For analyses not intending to use the power of double/triple scans,
or for data which will be widened and scanner type/site is not of
interest.
Only subject timepoints that have several entries (i.e. several scan sites),
will be reduced. All timepoints will be retained.
}
\examples{
\dontrun{
site_keeper(MOAS)
site_keeper(MOAS, 'ousSkyra')
site_keeper(MOAS, 'ousAvanto')
}

}
