% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_keeper.R
\name{site_keeper}
\alias{site_keeper}
\title{Utility function for reducing double/triple scans to a single row}
\usage{
site_keeper(data, keep = "long", quiet = F)
}
\arguments{
\item{data}{The MOAS or a MOAS generated file.}

\item{keep}{A string specifying which data from double/triple scans to keep.}

\item{quiet}{logical, TRUE sets it to verbose
Available options are:
'long' - keep data from scanner with most data (default),
'ousAvanto' = keep 'ousAvanto',
'ousSkyra' = keep 'ousSkyra', or
'ousPrisma' = keep 'ousPrisma'.}
}
\value{
A MOAS type file with one line per subject and timepoint.
}
\description{
\code{site_keeper} returns a MOAS data.frame with one row per
participant and timepoint (i.e. removes double/triple scan entries).
For analyses not intending to use the power of double/triple scans,
or for data which will be widened and scanner type/site is not of
interest.
Only subject timepoints that have several entries (i.e. several scan sites),
will be reduced. All timepoints will be retained.
}
\examples{
site_keeper(MOAS)
site_keeper(MOAS, 'ousSkyra')
site_keeper(MOAS, 'ousAvanto')

}
