% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{site_keeper}
\alias{site_keeper}
\title{Deprecated, use filter_site}
\usage{
site_keeper(...)
}
\arguments{
\item{...}{any argument to filter_site}
}
\value{
A MOAS type file with one line per subject and timepoint.
}
\description{
[\code{filter_site}] returns a MOAS data.frame with one row per
participant and timepoint (i.e. removes double/triple scan entries).
For analyses not intending to use the power of double/triple scans,
or for data which will be widened and scanner type/site is not of
interest.
Only subject timepoints that have several entries (i.e. several scan sites),
will be reduced. All timepoints will be retained.
}
\details{
Available options for 'keep' are:
\itemize{
'long' - keep data from scanner with most data (default),
'ousAvanto' = keep 'ousAvanto',
'ousSkyra' = keep 'ousSkyra', or
'ousPrisma' = keep 'ousPrisma'.
}

Available options for 'tie' are:
\itemize{
'interval' - keep data from scanner with longest data interval (default),
'ousAvanto' = keep 'ousAvanto',
'ousSkyra' = keep 'ousSkyra', or
'ousPrisma' = keep 'ousPrisma'.
}
}
\examples{
\dontrun{
filter_site(MOAS)
filter_site(MOAS, 'ousSkyra')
filter_site(MOAS, 'ousAvanto')
}

}
