% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widen.R
\name{widen}
\alias{widen}
\title{Widen the MOAS data.frame}
\usage{
widen(data, by, keep = NA)
}
\arguments{
\item{data}{The MOAS or a MOAS generated file.}

\item{by}{Column which to widen by. Options are: "Subject_Timepoint","Project_Wave","Site_Name", or "Site_Number".}
}
\value{
A MOAS type file widened by dplyr::selected column, prefixed with columns specifications
}
\description{
\code{widen} returns a widened data.frame by the columns specified
}
\details{
This is a function to create a widened data.frame from the MOAS.
Widening of the data is necessary for performing ANOVAs and
other statistical analysis, particularly if intending to run
in SPSS.

For MOAS which inclused double/triple scans, by="Site_Name",
must be run prior to by="Project_Wave".
}
\examples{
widen(MOAS, by="Site_Name")
widen(MOAS, by="Project_Wave")

# Data with double/triple scans must be widened in a two-stage process
dt = widen(MOAS, by="Site_Name")
df = widen(dt, by="Project_Wave")

}
